Исправлены предупреждения в TestWindow и MainWindow

+++ mainwindow.cpp
@@ -1601,7 +1601,7 @@ void MainWindow::on_leftBlockBCMButton_clicked()
 		selectBlockVersionComboBox->clear();
 		for (int i = 0; i < blockVersionsBCM.size(); i++)
 			selectBlockVersionComboBox->addItem(blockVersionsBCM[i]);
-		selectBlockVersionComboBox->setCurrentIndex(blockVersionsBCM.size() - 1);
+		selectBlockVersionComboBox->setCurrentIndex((int)blockVersionsBCM.size() - 1);
 		switchStyleMainButtons();
 	}
 }
@@ -1616,7 +1616,7 @@ void MainWindow::on_leftBlockDMButton_clicked()
 		selectBlockVersionComboBox->clear();
 		for (int i = 0; i < blockVersionsDTM.size(); i++)
 			selectBlockVersionComboBox->addItem(blockVersionsDTM[i]);
-		selectBlockVersionComboBox->setCurrentIndex(blockVersionsDTM.size() - 1);
+		selectBlockVersionComboBox->setCurrentIndex((int)blockVersionsDTM.size() - 1);
 		switchStyleMainButtons();
 	}
 }
